+++ src/providers/types/Firefly.ts
@@ -340,6 +340,9 @@ export type MetricsDownloadResponse = Response<{
 export type MetricsUploadResponse = Response<void>;
 
 export type BindResponse = Response<{
+    fid: string;
+    farcaster_signer_public_key?: string;
+    farcaster_signer_private_key?: string;
     account_id: string;
     account_raw_id: number;
     twitters: Array<{
