+++ src/providers/warpcast/createAccountByRelayService.ts
@@ -6,7 +6,7 @@ import { parseUrl } from '@/helpers/parseUrl.js';
 import { FarcasterSession } from '@/providers/farcaster/Session.js';
 import { FarcasterSocialMediaProvider } from '@/providers/farcaster/SocialMedia.js';
 import type { Account } from '@/providers/types/Account.js';
-import { bindOrRestoreFireflySession } from '@/services/bindOrRestoreFireflySession.js';
+import { bindOrRestoreFireflySession } from '@/services/bindFireflySession.js';
 
 interface FarcasterReplyResponse {
     channelToken: string;
