+++ src/services/bindFireflySession.ts
@@ -11,6 +11,7 @@ import { TwitterSession } from '@/providers/twitter/Session.js';
 import type { BindResponse } from '@/providers/types/Firefly.js';
 import type { Session } from '@/providers/types/Session.js';
 import { SessionType } from '@/providers/types/SocialMedia.js';
+import { restoreFireflySession } from '@/services/restoreFireflySession.js';
 import { settings } from '@/settings/index.js';
 import type { ResponseJSON } from '@/types/index.js';
 
@@ -87,7 +88,7 @@ async function bindTwitterSessionToFirefly(session: TwitterSession, signal?: Abo
  * @param signal
  * @returns
  */
-export async function bindFireflySession(session: Session, signal?: AbortSignal) {
+async function bindFireflySession(session: Session, signal?: AbortSignal) {
     // Ensure that the Firefly session is resumed before calling this function.
     fireflySessionHolder.assertSession();
 
@@ -107,5 +108,30 @@ export async function bindFireflySession(session: Session, signal?: AbortSignal)
         default:
             safeUnreachable(session.type);
             throw new UnreachableError('[bindFireflySession] session type', session.type);
+    }
+}
+
+export async function bindOrRestoreFireflySession(session: Session, signal?: AbortSignal) {
+    try {
+        const farcasterSession = session as FarcasterSession;
+        if (FarcasterSession.isCustodyWallet(farcasterSession)) throw new NotAllowedError();
+
+        if (fireflySessionHolder.session) {
+            const response = await bindFireflySession(session, signal);
+
+            if (FarcasterSession.isRelayService(session)) {
+                session.profileId = response.fid;
+            }
+
+            // this will return the existing session
+            return fireflySessionHolder.assertSession();
+        } else {
+            throw new Error('Firefly session is not available');
+        }
+    } catch (error) {
+        console.error(`[bindOrRestoreFireflySession] failed to bind firefly session ${error}`);
+
+        // this will create a new session
+        return restoreFireflySession(session, signal);
     }
 }
